[CI] Shorten build-docs run time (#565)
Currently, the CI process takes nearly 3 hours to complete the
`build-docs` step. We suspect this is related to the recent change in
the `matmul` backend from `simt` to `mma`, which significantly expanded
the tuning space. To address this issue, we are adjusting the precision
from `float32` to `float16`. This change reduces the tuning complexity
and is expected to decrease compilation time, helping to optimize the CI
workflow. After this PR, the running time for `build-docs` should fall
back to 40 minutes.

+++ gallery/getting-started/quick-start.py
@@ -33,9 +33,9 @@ import hidet
 import torch
 
 # take resnet18 as an example
-x = torch.randn(1, 3, 224, 224).cuda()
+x = torch.randn(1, 3, 224, 224, dtype=torch.float16).cuda()
 model = torch.hub.load('pytorch/vision:v0.9.0', 'resnet18', pretrained=True, verbose=False)
-model = model.cuda().eval()
+model = model.cuda().eval().to(torch.float16)
 
 # optimize the model with 'hidet' backend
 model_opt = torch.compile(model, backend='hidet', mode='max-autotune')
